Move LocalDevice tests to the right file
